+++ common/src/main/java/org/exodusstudio/stellaris/common/items/tools/CustomAxeItem.java
@@ -1,10 +1,10 @@
 package org.exodusstudio.stellaris.common.items.tools;
 
-import net.minecraft.world.item.Item;
+import net.minecraft.world.item.AxeItem;
 import net.minecraft.world.item.ToolMaterial;
 
-public class CustomAxeItem extends Item {
+public class CustomAxeItem extends AxeItem {
     public CustomAxeItem(Properties properties, ToolMaterial material, float attackDamage, float attackSpeed) {
-        super(properties.axe(material, attackDamage, attackSpeed));
+        super(material, attackDamage, attackSpeed, properties);
     }
 }