+++ common/src/main/java/org/exodusstudio/stellaris/common/items/tools/CustomHoeItem.java
@@ -1,10 +1,10 @@
 package org.exodusstudio.stellaris.common.items.tools;
 
-import net.minecraft.world.item.Item;
+import net.minecraft.world.item.HoeItem;
 import net.minecraft.world.item.ToolMaterial;
 
-public class CustomHoeItem extends Item {
+public class CustomHoeItem extends HoeItem {
     public CustomHoeItem(Properties properties, ToolMaterial material, float attackDamage, float attackSpeed) {
-        super(properties.hoe(material, attackDamage, attackSpeed));
+        super(material, attackDamage, attackSpeed, properties);
     }
 }