+++ common/src/main/java/org/exodusstudio/stellaris/common/items/tools/CustomShovelItem.java
@@ -1,10 +1,10 @@
 package org.exodusstudio.stellaris.common.items.tools;
 
-import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ShovelItem;
 import net.minecraft.world.item.ToolMaterial;
 
-public class CustomShovelItem extends Item {
+public class CustomShovelItem extends ShovelItem {
     public CustomShovelItem(Properties properties, ToolMaterial material, float attackDamage, float attackSpeed) {
-        super(properties.shovel(material, attackDamage, attackSpeed));
+        super(material, attackDamage, attackSpeed, properties);
     }
 }